Remove SerializableClosure security provider
If a secret key was set on the SerializableClosure then any future closure will be signed. This occurs in the Laravel framework. The child processes are not ran in Laravel thus not using the secret key that was set thus failing. This edit will temporarily remove the security provider, create an unsigned closure and set the provider back to original.