Fail if `unserialize` is unable to parse output
In `ParallelProcess` in `getOutput()` is checked if the `unserialize` is successful. 
When it fails, is sets the `errorOutput` to the process output. 
But this is not validated anymore because the check is already passed.
Moving the `getOutput()` before the error output check will ensure all error output cases are handled properly.